requirements.txt
agno==1.5.6
openai==1.82.1
google-search-results==2.4.2
yfinance==0.2.61
python-dotenv==1.1.0
